Struct google-games1::AchievementMethodsBuilder

pub struct AchievementMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on achievement resources. It is not used directly, but through the Games hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-games1" as games1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use games1::Games;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Games::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `increment(...)`, `list(...)`, `reveal(...)`, `set_steps_at_least(...)`, `unlock(...)` and `update_multiple(...)`
// to build up your call.
let rb = hub.achievements();

Methods

impl<'a, C, NC, A> AchievementMethodsBuilder<'a, C, NC, A>

fn list(&self, player_id: &str) -> AchievementListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists the progress for all your application's achievements for the currently authenticated player.

fn increment(&self, achievement_id: &str, steps_to_increment: i32) -> AchievementIncrementMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Increments the steps of the achievement with the given ID for the currently authenticated player.

fn set_steps_at_least(&self, achievement_id: &str, steps: i32) -> AchievementSetStepsAtLeastMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sets the steps for the currently authenticated player towards unlocking an achievement. If the steps parameter is less than the current number of steps that the player already gained for the achievement, the achievement is not modified.

fn update_multiple(&self, request: &AchievementUpdateMultipleRequest) -> AchievementUpdateMultipleMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates multiple achievements for the currently authenticated player.

fn reveal(&self, achievement_id: &str) -> AchievementRevealMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sets the state of the achievement with the given ID to REVEALED for the currently authenticated player.

fn unlock(&self, achievement_id: &str) -> AchievementUnlockMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Unlocks this achievement for the currently authenticated player.