Struct google_games1::ApplicationMethods

pub struct ApplicationMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on application resources. It is not used directly, but through the Games hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_games1 as games1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use games1::Games;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Games::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `played(...)` and `verify(...)`
// to build up your call.
let rb = hub.applications();

Methods

impl<'a, C, A> ApplicationMethods<'a, C, A>

fn verify(&self, application_id: &str) -> ApplicationVerifyCall<'a, C, A>

Create a builder to help you perform the following task:

Verifies the auth token provided with this request is for the application with the specified ID, and returns the ID of the player it was granted for.

Arguments

applicationId - The application ID from the Google Play developer console.

fn get(&self, application_id: &str) -> ApplicationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the metadata of the application with the given ID. If the requested application is not available for the specified platformType, the returned response will not include any instance data.

Arguments

applicationId - The application ID from the Google Play developer console.

fn played(&self) -> ApplicationPlayedCall<'a, C, A>

Create a builder to help you perform the following task:

Indicate that the the currently authenticated user is playing your application.