Struct google_games1::EventBatchRecordFailure

pub struct EventBatchRecordFailure {
    pub kind: Option<String>,
    pub range: Option<EventPeriodRange>,
    pub failure_cause: Option<String>,
}

This is a JSON template for a batch update failure resource.

This type is not used in any activity, and only used as part of another schema.

Fields

kind

Uniquely identifies the type of this resource. Value is always the fixed string games#eventBatchRecordFailure.

range

The time range which was rejected; empty for a request-wide failure.

failure_cause

The cause for the update failure. Possible values are:
- "TOO_LARGE": A batch request was issued with more events than are allowed in a single batch. - "TIME_PERIOD_EXPIRED": A batch was sent with data too far in the past to record. - "TIME_PERIOD_SHORT": A batch was sent with a time range that was too short. - "TIME_PERIOD_LONG": A batch was sent with a time range that was too long. - "ALREADY_UPDATED": An attempt was made to record a batch of data which was already seen. - "RECORD_RATE_HIGH": An attempt was made to record data faster than the server will apply updates.