Struct google_games1::Games

pub struct Games<C, A> {
    // some fields omitted
}

Central instance to access all Games related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_games1 as games1;
use games1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use games1::Games;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Games::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.turn_based_matches().leave_turn("matchId", -18)
             .pending_participant_id("kasd")
             .language("accusam")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Methods

impl<'a, C, A> Games<C, A> where C: BorrowMut<Client>, A: GetToken

fn new(client: C, authenticator: A) -> Games<C, A>

fn achievement_definitions(&'a self) -> AchievementDefinitionMethods<'a, C, A>

fn achievements(&'a self) -> AchievementMethods<'a, C, A>

fn applications(&'a self) -> ApplicationMethods<'a, C, A>

fn events(&'a self) -> EventMethods<'a, C, A>

fn leaderboards(&'a self) -> LeaderboardMethods<'a, C, A>

fn metagame(&'a self) -> MetagameMethods<'a, C, A>

fn players(&'a self) -> PlayerMethods<'a, C, A>

fn pushtokens(&'a self) -> PushtokenMethods<'a, C, A>

fn quest_milestones(&'a self) -> QuestMilestoneMethods<'a, C, A>

fn quests(&'a self) -> QuestMethods<'a, C, A>

fn revisions(&'a self) -> RevisionMethods<'a, C, A>

fn rooms(&'a self) -> RoomMethods<'a, C, A>

fn scores(&'a self) -> ScoreMethods<'a, C, A>

fn snapshots(&'a self) -> SnapshotMethods<'a, C, A>

fn turn_based_matches(&'a self) -> TurnBasedMatcheMethods<'a, C, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.1.8.

Returns the previously set user-agent.