Struct google_games1::Instance

pub struct Instance {
    pub kind: Option<String>,
    pub acquisition_uri: Option<String>,
    pub name: Option<String>,
    pub turn_based_play: Option<bool>,
    pub web_instance: Option<InstanceWebDetails>,
    pub android_instance: Option<InstanceAndroidDetails>,
    pub ios_instance: Option<InstanceIosDetails>,
    pub platform_type: Option<String>,
    pub realtime_play: Option<bool>,
}

This is a JSON template for the Instance resource.

This type is not used in any activity, and only used as part of another schema.

Fields
kind: Option<String>

Uniquely identifies the type of this resource. Value is always the fixed string games#instance.

acquisition_uri: Option<String>

URI which shows where a user can acquire this instance.

name: Option<String>

Localized display name.

turn_based_play: Option<bool>

Flag to show if this game instance supports turn based play.

web_instance: Option<InstanceWebDetails>

Platform dependent details for Web.

android_instance: Option<InstanceAndroidDetails>

Platform dependent details for Android.

ios_instance: Option<InstanceIosDetails>

Platform dependent details for iOS.

platform_type: Option<String>

The platform type. Possible values are:
- "ANDROID" - Instance is for Android. - "IOS" - Instance is for iOS - "WEB_APP" - Instance is for Web App.

realtime_play: Option<bool>

Flag to show if this game instance supports realtime play.