Struct google_games1::Player

pub struct Player {
    pub original_player_id: Option<String>,
    pub kind: Option<String>,
    pub display_name: Option<String>,
    pub name: Option<PlayerName>,
    pub last_played_with: Option<Played>,
    pub player_id: Option<String>,
    pub banner_url_portrait: Option<String>,
    pub title: Option<String>,
    pub experience_info: Option<PlayerExperienceInfo>,
    pub banner_url_landscape: Option<String>,
    pub avatar_image_url: Option<String>,
}

This is a JSON template for a Player resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list players (none)

get players (response)

Fields

original_player_id

The player ID that was used for this player the first time they signed into the game in question. This is only populated for calls to player.get for the requesting player, only if the player ID has subsequently changed, and only to clients that support remapping player IDs.

kind

Uniquely identifies the type of this resource. Value is always the fixed string games#player.

display_name

The name to display for the player.

name

An object representation of the individual components of the player's name. For some players, these fields may not be present.

last_played_with

Details about the last time this player played a multiplayer game with the currently authenticated player. Populated for PLAYED_WITH player collection members.

player_id

The ID of the player.

banner_url_portrait

The url to the portrait mode player banner image.

title

The player's title rewarded for their game activities.

experience_info

An object to represent Play Game experience information for the player.

banner_url_landscape

The url to the landscape mode player banner image.

avatar_image_url

The base URL for the image that represents the player.