Struct google_games1::PlayerLeaderboardScore

pub struct PlayerLeaderboardScore {
    pub write_timestamp: Option<String>,
    pub kind: Option<String>,
    pub public_rank: Option<LeaderboardScoreRank>,
    pub score_tag: Option<String>,
    pub time_span: Option<String>,
    pub score_string: Option<String>,
    pub leaderboard_id: Option<String>,
    pub score_value: Option<String>,
    pub social_rank: Option<LeaderboardScoreRank>,
}

This is a JSON template for a player leaderboard score object.

This type is not used in any activity, and only used as part of another schema.

Fields

write_timestamp

The timestamp at which this score was recorded, in milliseconds since the epoch in UTC.

kind

Uniquely identifies the type of this resource. Value is always the fixed string games#playerLeaderboardScore.

public_rank

The public rank of the score in this leaderboard. This object will not be present if the user is not sharing their scores publicly.

score_tag

Additional information about the score. Values must contain no more than 64 URI-safe characters as defined by section 2.3 of RFC 3986.

time_span

The time span of this score. Possible values are:
- "ALL_TIME" - The score is an all-time score. - "WEEKLY" - The score is a weekly score. - "DAILY" - The score is a daily score.

score_string

The formatted value of this score.

leaderboard_id

The ID of the leaderboard this score is in.

score_value

The numerical value of this score.

social_rank

The social rank of the score in this leaderboard.