Struct google_games1::PlayerMethods

pub struct PlayerMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on player resources. It is not used directly, but through the Games hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_games1 as games1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use games1::Games;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Games::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)` and `list(...)`
// to build up your call.
let rb = hub.players();

Methods

impl<'a, C, A> PlayerMethods<'a, C, A>

fn list(&self, collection: &str) -> PlayerListCall<'a, C, A>

Create a builder to help you perform the following task:

Get the collection of players for the currently authenticated user.

Arguments

collection - Collection of players being retrieved

fn get(&self, player_id: &str) -> PlayerGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the Player resource with the given ID. To retrieve the player for the currently authenticated user, set playerId to me.

Arguments

playerId - A player ID. A value of me may be used in place of the authenticated player's ID.