Struct google_games1::QuestMethods

pub struct QuestMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on quest resources. It is not used directly, but through the Games hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_games1 as games1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use games1::Games;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Games::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `accept(...)` and `list(...)`
// to build up your call.
let rb = hub.quests();

Methods

impl<'a, C, A> QuestMethods<'a, C, A>

fn accept(&self, quest_id: &str) -> QuestAcceptCall<'a, C, A>

Create a builder to help you perform the following task:

Indicates that the currently authorized user will participate in the quest.

Arguments

questId - The ID of the quest.

fn list(&self, player_id: &str) -> QuestListCall<'a, C, A>

Create a builder to help you perform the following task:

Get a list of quests for your application and the currently authenticated player.

Arguments

playerId - A player ID. A value of me may be used in place of the authenticated player's ID.