Struct google-games1::RoomListMethodBuilder

pub struct RoomListMethodBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

Returns invitations to join rooms.

A builder for the list method supported by a room resource. It is not used directly, but through a RoomMethodsBuilder.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.rooms().list()
             .page_token("takimata")
             .max_results(-27)
             .language("labore")
             .doit();

Methods

impl<'a, C, NC, A> RoomListMethodBuilder<'a, C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>>, A: GetToken

fn doit(self) -> Result<(Response, RoomList)>

Perform the operation you have build so far.

fn page_token(self, new_value: &str) -> RoomListMethodBuilder<'a, C, NC, A>

Sets the page token query property to the given value.

The token returned by the previous request.

fn max_results(self, new_value: i32) -> RoomListMethodBuilder<'a, C, NC, A>

Sets the max results query property to the given value.

The maximum number of rooms to return in the response, used for paging. For any response, the actual number of rooms to return may be less than the specified maxResults.

fn language(self, new_value: &str) -> RoomListMethodBuilder<'a, C, NC, A>

Sets the language query property to the given value.

The preferred language to use for strings returned by this method.

fn delegate(self, new_value: &'a mut Delegate) -> RoomListMethodBuilder<'a, C, NC, A>

Sets the delegate property to the given value.

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

fn param<T>(self, name: T, value: T) -> RoomListMethodBuilder<'a, C, NC, A> where T: Str

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> RoomListMethodBuilder<'a, C, NC, A> where T: Str

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead of relying on the automated algorithm which simply prefers read-only scopes over those who are not.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.