Struct google-games1::RoomMethodsBuilder

pub struct RoomMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on room resources. It is not used directly, but through the Games hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-games1" as games1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use games1::Games;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Games::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `create(...)`, `decline(...)`, `dismiss(...)`, `get(...)`, `join(...)`, `leave(...)`, `list(...)` and `report_status(...)`
// to build up your call.
let rb = hub.rooms();

Methods

impl<'a, C, NC, A> RoomMethodsBuilder<'a, C, NC, A>

fn get(&self, room_id: &str) -> RoomGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get the data for a room.

fn leave(&self, request: &RoomLeaveRequest, room_id: &str) -> RoomLeaveMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Leave a room. For internal use by the Games SDK only. Calling this method directly is unsupported.

fn list(&self) -> RoomListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns invitations to join rooms.

fn report_status(&self, request: &RoomP2PStatuses, room_id: &str) -> RoomReportStatuMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates sent by a client reporting the status of peers in a room. For internal use by the Games SDK only. Calling this method directly is unsupported.

fn create(&self, request: &RoomCreateRequest) -> RoomCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a room. For internal use by the Games SDK only. Calling this method directly is unsupported.

fn decline(&self, room_id: &str) -> RoomDeclineMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Decline an invitation to join a room. For internal use by the Games SDK only. Calling this method directly is unsupported.

fn dismiss(&self, room_id: &str) -> RoomDismisMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Dismiss an invitation to join a room. For internal use by the Games SDK only. Calling this method directly is unsupported.

fn join(&self, request: &RoomJoinRequest, room_id: &str) -> RoomJoinMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Join a room. For internal use by the Games SDK only. Calling this method directly is unsupported.