Struct google_games1::RoomParticipant

pub struct RoomParticipant {
    pub auto_matched: Option<bool>,
    pub status: Option<String>,
    pub kind: Option<String>,
    pub auto_matched_player: Option<AnonymousPlayer>,
    pub client_address: Option<RoomClientAddress>,
    pub capabilities: Option<Vec<String>>,
    pub player: Option<Player>,
    pub leave_reason: Option<String>,
    pub connected: Option<bool>,
    pub id: Option<String>,
}

This is a JSON template for a participant in a room.

This type is not used in any activity, and only used as part of another schema.

Fields
auto_matched: Option<bool>

True if this participant was auto-matched with the requesting player.

status: Option<String>

The status of the participant with respect to the room. Possible values are:
- "PARTICIPANT_INVITED" - The participant has been invited to join the room, but has not yet responded. - "PARTICIPANT_JOINED" - The participant has joined the room (either after creating it or accepting an invitation.) - "PARTICIPANT_DECLINED" - The participant declined an invitation to join the room. - "PARTICIPANT_LEFT" - The participant joined the room and then left it.

kind: Option<String>

Uniquely identifies the type of this resource. Value is always the fixed string games#roomParticipant.

auto_matched_player: Option<AnonymousPlayer>

Information about a player that has been anonymously auto-matched against the requesting player. (Either player or autoMatchedPlayer will be set.)

client_address: Option<RoomClientAddress>

Client address for the participant.

capabilities: Option<Vec<String>>

The capabilities which can be used when communicating with this participant.

player: Option<Player>

Information about the player. Not populated if this player was anonymously auto-matched against the requesting player. (Either player or autoMatchedPlayer will be set.)

leave_reason: Option<String>

The reason the participant left the room; populated if the participant status is PARTICIPANT_LEFT. Possible values are:
- "PLAYER_LEFT" - The player explicitly chose to leave the room. - "GAME_LEFT" - The game chose to remove the player from the room. - "ABANDONED" - The player switched to another application and abandoned the room. - "PEER_CONNECTION_FAILURE" - The client was unable to establish or maintain a connection to other peer(s) in the room. - "SERVER_ERROR" - The client received an error response when it tried to communicate with the server. - "TIMEOUT" - The client timed out while waiting for players to join and connect. - "PRESENCE_FAILURE" - The client's XMPP connection ended abruptly.

connected: Option<bool>

True if this participant is in the fully connected set of peers in the room.

id: Option<String>

An identifier for the participant in the scope of the room. Cannot be used to identify a player across rooms or in other contexts.