Struct google_games1::RoomStatus

pub struct RoomStatus {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub status_version: Option<i32>,
    pub room_id: Option<String>,
    pub auto_matching_status: Option<RoomAutoMatchStatus>,
    pub participants: Option<Vec<RoomParticipant>>,
}

This is a JSON template for the status of a room that the player has joined.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

report status rooms (response)

Fields

status

The status of the room. Possible values are:
- "ROOM_INVITING" - One or more players have been invited and not responded. - "ROOM_AUTO_MATCHING" - One or more slots need to be filled by auto-matching. - "ROOM_CONNECTING" - Players have joined are connecting to each other (either before or after auto-matching). - "ROOM_ACTIVE" - All players have joined and connected to each other. - "ROOM_DELETED" - All joined players have left.

kind

Uniquely identifies the type of this resource. Value is always the fixed string games#roomStatus.

status_version

The version of the status for the room: an increasing counter, used by the client to ignore out-of-order updates to room status.

room_id

Globally unique ID for a room.

auto_matching_status

Auto-matching status for this room. Not set if the room is not currently in the automatching queue.

participants

The participants involved in the room, along with their statuses. Includes participants who have left or declined invitations.