Struct google_games1::ScoreGetCall

pub struct ScoreGetCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Get high scores, and optionally ranks, in leaderboards for the currently authenticated player. For a specific time span, leaderboardId can be set to ALL to retrieve data for all leaderboards in a given time span. NOTE: You cannot ask for 'ALL' leaderboards and 'ALL' timeSpans in the same request; only one parameter may be set to 'ALL'.

A builder for the get method supported by a score resource. It is not used directly, but through a ScoreMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.scores().get("playerId", "leaderboardId", "timeSpan")
             .page_token("sea")
             .max_results(-85)
             .language("ipsum")
             .include_rank_type("aliquyam")
             .doit();

Methods

impl<'a, C, A> ScoreGetCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, PlayerLeaderboardScoreListResponse)>

Perform the operation you have build so far.

fn player_id(self, new_value: &str) -> ScoreGetCall<'a, C, A>

A player ID. A value of me may be used in place of the authenticated player's ID.

Sets the player id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn leaderboard_id(self, new_value: &str) -> ScoreGetCall<'a, C, A>

The ID of the leaderboard. Can be set to 'ALL' to retrieve data for all leaderboards for this application.

Sets the leaderboard id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn time_span(self, new_value: &str) -> ScoreGetCall<'a, C, A>

The time span for the scores and ranks you're requesting.

Sets the time span path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn page_token(self, new_value: &str) -> ScoreGetCall<'a, C, A>

The token returned by the previous request.

Sets the page token query property to the given value.

fn max_results(self, new_value: i32) -> ScoreGetCall<'a, C, A>

The maximum number of leaderboard scores to return in the response. For any response, the actual number of leaderboard scores returned may be less than the specified maxResults.

Sets the max results query property to the given value.

fn language(self, new_value: &str) -> ScoreGetCall<'a, C, A>

The preferred language to use for strings returned by this method.

Sets the language query property to the given value.

fn include_rank_type(self, new_value: &str) -> ScoreGetCall<'a, C, A>

The types of ranks to return. If the parameter is omitted, no ranks will be returned.

Sets the include rank type query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> ScoreGetCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> ScoreGetCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> ScoreGetCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Full.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.