Struct google_games1::ScoreSubmission

pub struct ScoreSubmission {
    pub kind: Option<String>,
    pub leaderboard_id: Option<String>,
    pub score: Option<String>,
    pub score_tag: Option<String>,
    pub signature: Option<String>,
}

This is a JSON template for a request to submit a score to leaderboards.

This type is not used in any activity, and only used as part of another schema.

Fields
kind: Option<String>

Uniquely identifies the type of this resource. Value is always the fixed string games#scoreSubmission.

leaderboard_id: Option<String>

The leaderboard this score is being submitted to.

score: Option<String>

The new score being submitted.

score_tag: Option<String>

Additional information about this score. Values will contain no more than 64 URI-safe characters as defined by section 2.3 of RFC 3986.

signature: Option<String>

Signature Values will contain URI-safe characters as defined by section 2.3 of RFC 3986.