Struct google_games1::Snapshot

pub struct Snapshot {
    pub kind: Option<String>,
    pub description: Option<String>,
    pub title: Option<String>,
    pub cover_image: Option<SnapshotImage>,
    pub last_modified_millis: Option<String>,
    pub id: Option<String>,
    pub drive_id: Option<String>,
    pub duration_millis: Option<String>,
    pub unique_name: Option<String>,
    pub type_: Option<String>,
    pub progress_value: Option<String>,
}

This is a JSON template for an snapshot object.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list snapshots (none)

get snapshots (response)

Fields
kind: Option<String>

Uniquely identifies the type of this resource. Value is always the fixed string games#snapshot.

description: Option<String>

The description of this snapshot.

title: Option<String>

The title of this snapshot.

cover_image: Option<SnapshotImage>

The cover image of this snapshot. May be absent if there is no image.

last_modified_millis: Option<String>

The timestamp (in millis since Unix epoch) of the last modification to this snapshot.

id: Option<String>

The ID of the snapshot.

drive_id: Option<String>

The ID of the file underlying this snapshot in the Drive API. Only present if the snapshot is a view on a Drive file and the file is owned by the caller.

duration_millis: Option<String>

The duration associated with this snapshot, in millis.

unique_name: Option<String>

The unique name provided when the snapshot was created.

type_: Option<String>

The type of this snapshot. Possible values are:
- "SAVE_GAME" - A snapshot representing a save game.

progress_value: Option<String>

The progress value (64-bit integer set by developer) associated with this snapshot.