Struct google_games1::SnapshotMethods

pub struct SnapshotMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on snapshot resources. It is not used directly, but through the Games hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_games1 as games1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use games1::Games;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Games::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)` and `list(...)`
// to build up your call.
let rb = hub.snapshots();

Methods

impl<'a, C, A> SnapshotMethods<'a, C, A>

fn get(&self, snapshot_id: &str) -> SnapshotGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the metadata for a given snapshot ID.

Arguments

snapshotId - The ID of the snapshot.

fn list(&self, player_id: &str) -> SnapshotListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of snapshots created by your application for the player corresponding to the player ID.

Arguments

playerId - A player ID. A value of me may be used in place of the authenticated player's ID.