Struct google_games1::TurnBasedMatch

pub struct TurnBasedMatch {
    pub status: Option<String>,
    pub auto_matching_criteria: Option<TurnBasedAutoMatchingCriteria>,
    pub inviter_id: Option<String>,
    pub match_version: Option<i32>,
    pub variant: Option<i32>,
    pub match_id: Option<String>,
    pub rematch_id: Option<String>,
    pub results: Option<Vec<ParticipantResult>>,
    pub match_number: Option<i32>,
    pub previous_match_data: Option<TurnBasedMatchData>,
    pub application_id: Option<String>,
    pub description: Option<String>,
    pub with_participant_id: Option<String>,
    pub kind: Option<String>,
    pub creation_details: Option<TurnBasedMatchModification>,
    pub user_match_status: Option<String>,
    pub data: Option<TurnBasedMatchData>,
    pub participants: Option<Vec<TurnBasedMatchParticipant>>,
    pub pending_participant_id: Option<String>,
    pub last_update_details: Option<TurnBasedMatchModification>,
}

This is a JSON template for a turn-based match resource object.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

decline turn based matches (response)

leave turn based matches (response)

finish turn based matches (response)

take turn turn based matches (response)

create turn based matches (response)

join turn based matches (response)

leave turn turn based matches (response)

get turn based matches (response)

Fields

status

The status of the match. Possible values are:
- "MATCH_AUTO_MATCHING" - One or more slots need to be filled by auto-matching; the match cannot be established until they are filled. - "MATCH_ACTIVE" - The match has started. - "MATCH_COMPLETE" - The match has finished. - "MATCH_CANCELED" - The match was canceled. - "MATCH_EXPIRED" - The match expired due to inactivity. - "MATCH_DELETED" - The match should no longer be shown on the client. Returned only for tombstones for matches when sync is called.

auto_matching_criteria

Criteria for auto-matching players into this match.

inviter_id

The ID of the participant that invited the user to the match. Not set if the user was not invited to the match.

match_version

The version of this match: an increasing counter, used to avoid out-of-date updates to the match.

variant

The variant / mode of the application being played; can be any integer value, or left blank.

match_id

Globally unique ID for a turn-based match.

rematch_id

The ID of a rematch of this match. Only set for completed matches that have been rematched.

results

The results reported for this match.

match_number

The number of the match in a chain of rematches. Will be set to 1 for the first match and incremented by 1 for each rematch.

previous_match_data

The data / game state for the previous match; set for the first turn of rematches only.

application_id

The ID of the application being played.

description

This short description is generated by our servers based on turn state and is localized and worded relative to the player requesting the match. It is intended to be displayed when the match is shown in a list.

with_participant_id

The ID of another participant in the match that can be used when describing the participants the user is playing with.

kind

Uniquely identifies the type of this resource. Value is always the fixed string games#turnBasedMatch.

creation_details

Details about the match creation.

user_match_status

The status of the current user in the match. Derived from the match type, match status, the user's participant status, and the pending participant for the match. Possible values are:
- "USER_INVITED" - The user has been invited to join the match and has not responded yet. - "USER_AWAITING_TURN" - The user is waiting for their turn. - "USER_TURN" - The user has an action to take in the match. - "USER_MATCH_COMPLETED" - The match has ended (it is completed, canceled, or expired.)

data

The data / game state for this match.

participants

The participants involved in the match, along with their statuses. Includes participants who have left or declined invitations.

pending_participant_id

The ID of the participant that is taking a turn.

last_update_details

Details about the last update to the match.