Struct google_games1::TurnBasedMatchParticipant

pub struct TurnBasedMatchParticipant {
    pub auto_matched: Option<bool>,
    pub status: Option<String>,
    pub kind: Option<String>,
    pub auto_matched_player: Option<AnonymousPlayer>,
    pub player: Option<Player>,
    pub id: Option<String>,
}

This is a JSON template for a participant in a turn-based match.

This type is not used in any activity, and only used as part of another schema.

Fields
auto_matched: Option<bool>

True if this participant was auto-matched with the requesting player.

status: Option<String>

The status of the participant with respect to the match. Possible values are:
- "PARTICIPANT_NOT_INVITED_YET" - The participant is slated to be invited to the match, but the invitation has not been sent; the invite will be sent when it becomes their turn. - "PARTICIPANT_INVITED" - The participant has been invited to join the match, but has not yet responded. - "PARTICIPANT_JOINED" - The participant has joined the match (either after creating it or accepting an invitation.) - "PARTICIPANT_DECLINED" - The participant declined an invitation to join the match. - "PARTICIPANT_LEFT" - The participant joined the match and then left it. - "PARTICIPANT_FINISHED" - The participant finished playing in the match. - "PARTICIPANT_UNRESPONSIVE" - The participant did not take their turn in the allotted time.

kind: Option<String>

Uniquely identifies the type of this resource. Value is always the fixed string games#turnBasedMatchParticipant.

auto_matched_player: Option<AnonymousPlayer>

Information about a player that has been anonymously auto-matched against the requesting player. (Either player or autoMatchedPlayer will be set.)

player: Option<Player>

Information about the player. Not populated if this player was anonymously auto-matched against the requesting player. (Either player or autoMatchedPlayer will be set.)

id: Option<String>

An identifier for the participant in the scope of the match. Cannot be used to identify a player across matches or in other contexts.