Struct google_games1::TurnBasedMatcheMethods

pub struct TurnBasedMatcheMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on turnBasedMatche resources. It is not used directly, but through the Games hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_games1 as games1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use games1::Games;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Games::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `cancel(...)`, `create(...)`, `decline(...)`, `dismiss(...)`, `finish(...)`, `get(...)`, `join(...)`, `leave(...)`, `leave_turn(...)`, `list(...)`, `rematch(...)`, `sync(...)` and `take_turn(...)`
// to build up your call.
let rb = hub.turn_based_matches();

Methods

impl<'a, C, A> TurnBasedMatcheMethods<'a, C, A>

fn dismiss(&self, match_id: &str) -> TurnBasedMatcheDismisCall<'a, C, A>

Create a builder to help you perform the following task:

Dismiss a turn-based match from the match list. The match will no longer show up in the list and will not generate notifications.

Arguments

matchId - The ID of the match.

fn sync(&self) -> TurnBasedMatcheSyncCall<'a, C, A>

Create a builder to help you perform the following task:

Returns turn-based matches the player is or was involved in that changed since the last sync call, with the least recent changes coming first. Matches that should be removed from the local cache will have a status of MATCH_DELETED.

fn decline(&self, match_id: &str) -> TurnBasedMatcheDeclineCall<'a, C, A>

Create a builder to help you perform the following task:

Decline an invitation to play a turn-based match.

Arguments

matchId - The ID of the match.

fn get(&self, match_id: &str) -> TurnBasedMatcheGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get the data for a turn-based match.

Arguments

matchId - The ID of the match.

fn create(&self, request: TurnBasedMatchCreateRequest) -> TurnBasedMatcheCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Create a turn-based match.

Arguments

request - No description provided.

fn join(&self, match_id: &str) -> TurnBasedMatcheJoinCall<'a, C, A>

Create a builder to help you perform the following task:

Join a turn-based match.

Arguments

matchId - The ID of the match.

fn leave_turn(&self, match_id: &str, match_version: i32) -> TurnBasedMatcheLeaveTurnCall<'a, C, A>

Create a builder to help you perform the following task:

Leave a turn-based match during the current player's turn, without canceling the match.

Arguments

matchId - The ID of the match.

matchVersion - The version of the match being updated.

fn cancel(&self, match_id: &str) -> TurnBasedMatcheCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Cancel a turn-based match.

Arguments

matchId - The ID of the match.

fn finish(&self, request: TurnBasedMatchResults, match_id: &str) -> TurnBasedMatcheFinishCall<'a, C, A>

Create a builder to help you perform the following task:

Finish a turn-based match. Each player should make this call once, after all results are in. Only the player whose turn it is may make the first call to Finish, and can pass in the final match state.

Arguments

request - No description provided.

matchId - The ID of the match.

fn leave(&self, match_id: &str) -> TurnBasedMatcheLeaveCall<'a, C, A>

Create a builder to help you perform the following task:

Leave a turn-based match when it is not the current player's turn, without canceling the match.

Arguments

matchId - The ID of the match.

fn list(&self) -> TurnBasedMatcheListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns turn-based matches the player is or was involved in.

fn take_turn(&self, request: TurnBasedMatchTurn, match_id: &str) -> TurnBasedMatcheTakeTurnCall<'a, C, A>

Create a builder to help you perform the following task:

Commit the results of a player turn.

Arguments

request - No description provided.

matchId - The ID of the match.

fn rematch(&self, match_id: &str) -> TurnBasedMatcheRematchCall<'a, C, A>

Create a builder to help you perform the following task:

Create a rematch of a match that was previously completed, with the same participants. This can be called by only one player on a match still in their list; the player must have called Finish first. Returns the newly created match; it will be the caller's turn.

Arguments

matchId - The ID of the match.