Struct google_games1::AchievementDefinition

pub struct AchievementDefinition {
    pub total_steps: Option<i32>,
    pub achievement_type: Option<String>,
    pub description: Option<String>,
    pub formatted_total_steps: Option<String>,
    pub kind: Option<String>,
    pub initial_state: Option<String>,
    pub experience_points: Option<String>,
    pub id: Option<String>,
    pub is_revealed_icon_url_default: Option<bool>,
    pub unlocked_icon_url: Option<String>,
    pub revealed_icon_url: Option<String>,
    pub is_unlocked_icon_url_default: Option<bool>,
    pub name: Option<String>,
}

This is a JSON template for an achievement definition object.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list achievement definitions (none)

Fields
total_steps: Option<i32>

The total steps for an incremental achievement.

achievement_type: Option<String>

The type of the achievement. Possible values are:
- "STANDARD" - Achievement is either locked or unlocked. - "INCREMENTAL" - Achievement is incremental.

description: Option<String>

The description of the achievement.

formatted_total_steps: Option<String>

The total steps for an incremental achievement as a string.

kind: Option<String>

Uniquely identifies the type of this resource. Value is always the fixed string games#achievementDefinition.

initial_state: Option<String>

The initial state of the achievement. Possible values are:
- "HIDDEN" - Achievement is hidden. - "REVEALED" - Achievement is revealed. - "UNLOCKED" - Achievement is unlocked.

experience_points: Option<String>

Experience points which will be earned when unlocking this achievement.

id: Option<String>

The ID of the achievement.

is_revealed_icon_url_default: Option<bool>

Indicates whether the revealed icon image being returned is a default image, or is provided by the game.

unlocked_icon_url: Option<String>

The image URL for the unlocked achievement icon.

revealed_icon_url: Option<String>

The image URL for the revealed achievement icon.

is_unlocked_icon_url_default: Option<bool>

Indicates whether the unlocked icon image being returned is a default image, or is game-provided.

name: Option<String>

The name of the achievement.