Struct google_games1::Application

pub struct Application {
    pub category: Option<ApplicationCategory>,
    pub theme_color: Option<String>,
    pub description: Option<String>,
    pub author: Option<String>,
    pub last_updated_timestamp: Option<String>,
    pub kind: Option<String>,
    pub instances: Option<Vec<Instance>>,
    pub enabled_features: Option<Vec<String>>,
    pub achievement_count: Option<i32>,
    pub leaderboard_count: Option<i32>,
    pub assets: Option<Vec<ImageAsset>>,
    pub id: Option<String>,
    pub name: Option<String>,
}

This is a JSON template for the Application resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get applications (response)

played applications (none)

Fields

category

The category of the application.

theme_color

A hint to the client UI for what color to use as an app-themed color. The color is given as an RGB triplet (e.g. "E0E0E0").

description

The description of the application.

author

The author of the application.

last_updated_timestamp

The last updated timestamp of the application.

kind

Uniquely identifies the type of this resource. Value is always the fixed string games#application.

instances

The instances of the application.

enabled_features

A list of features that have been enabled for the application. Possible values are:
- "SNAPSHOTS" - Snapshots has been enabled

achievement_count

The number of achievements visible to the currently authenticated player.

leaderboard_count

The number of leaderboards visible to the currently authenticated player.

assets

The assets of the application.

id

The ID of the application.

name

The name of the application.