Struct google_games1::Quest

pub struct Quest {
    pub description: Option<String>,
    pub banner_url: Option<String>,
    pub last_updated_timestamp_millis: Option<String>,
    pub accepted_timestamp_millis: Option<String>,
    pub icon_url: Option<String>,
    pub notify_timestamp_millis: Option<String>,
    pub application_id: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub milestones: Option<Vec<QuestMilestone>>,
    pub name: Option<String>,
    pub start_timestamp_millis: Option<String>,
    pub end_timestamp_millis: Option<String>,
    pub state: Option<String>,
    pub is_default_banner_url: Option<bool>,
    pub is_default_icon_url: Option<bool>,
}

This is a JSON template for a Quest resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

accept quests (response)

list quests (none)

Fields
description: Option<String>

The description of the quest.

banner_url: Option<String>

The banner image URL for the quest.

last_updated_timestamp_millis: Option<String>

The timestamp at which the quest was last updated by the user in milliseconds since the epoch in UTC. Only present if the player has accepted the quest.

accepted_timestamp_millis: Option<String>

The timestamp at which the user accepted the quest in milliseconds since the epoch in UTC. Only present if the player has accepted the quest.

icon_url: Option<String>

The icon image URL for the quest.

notify_timestamp_millis: Option<String>

The timestamp at which the user should be notified that the quest will end soon in milliseconds since the epoch in UTC.

application_id: Option<String>

The ID of the application this quest is part of.

id: Option<String>

The ID of the quest.

kind: Option<String>

Uniquely identifies the type of this resource. Value is always the fixed string games#quest.

milestones: Option<Vec<QuestMilestone>>

The quest milestones.

name: Option<String>

The name of the quest.

start_timestamp_millis: Option<String>

The timestamp at which the quest becomes active in milliseconds since the epoch in UTC.

end_timestamp_millis: Option<String>

The timestamp at which the quest ceases to be active in milliseconds since the epoch in UTC.

state: Option<String>

The state of the quest. Possible values are:
- "UPCOMING": The quest is upcoming. The user can see the quest, but cannot accept it until it is open. - "OPEN": The quest is currently open and may be accepted at this time. - "ACCEPTED": The user is currently participating in this quest. - "COMPLETED": The user has completed the quest. - "FAILED": The quest was attempted but was not completed before the deadline expired. - "EXPIRED": The quest has expired and was not accepted. - "DELETED": The quest should be deleted from the local database.

is_default_banner_url: Option<bool>

Indicates whether the banner image being returned is a default image, or is game-provided.

is_default_icon_url: Option<bool>

Indicates whether the icon image being returned is a default image, or is game-provided.