Struct google_games1::Room

pub struct Room {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub auto_matching_criteria: Option<RoomAutoMatchingCriteria>,
    pub creation_details: Option<RoomModification>,
    pub description: Option<String>,
    pub room_status_version: Option<i32>,
    pub auto_matching_status: Option<RoomAutoMatchStatus>,
    pub last_update_details: Option<RoomModification>,
    pub variant: Option<i32>,
    pub participants: Option<Vec<RoomParticipant>>,
    pub room_id: Option<String>,
    pub application_id: Option<String>,
    pub inviter_id: Option<String>,
}

This is a JSON template for a room resource object.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

decline rooms (response)

leave rooms (response)

dismiss rooms (none)

list rooms (none)

create rooms (response)

get rooms (response)

join rooms (response)

report status rooms (none)

Fields
status: Option<String>

The status of the room. Possible values are:
- "ROOM_INVITING" - One or more players have been invited and not responded. - "ROOM_AUTO_MATCHING" - One or more slots need to be filled by auto-matching. - "ROOM_CONNECTING" - Players have joined and are connecting to each other (either before or after auto-matching). - "ROOM_ACTIVE" - All players have joined and connected to each other. - "ROOM_DELETED" - The room should no longer be shown on the client. Returned in sync calls when a player joins a room (as a tombstone), or for rooms where all joined participants have left.

kind: Option<String>

Uniquely identifies the type of this resource. Value is always the fixed string games#room.

auto_matching_criteria: Option<RoomAutoMatchingCriteria>

Criteria for auto-matching players into this room.

creation_details: Option<RoomModification>

Details about the room creation.

description: Option<String>

This short description is generated by our servers and worded relative to the player requesting the room. It is intended to be displayed when the room is shown in a list (that is, an invitation to a room.)

room_status_version: Option<i32>

The version of the room status: an increasing counter, used by the client to ignore out-of-order updates to room status.

auto_matching_status: Option<RoomAutoMatchStatus>

Auto-matching status for this room. Not set if the room is not currently in the auto-matching queue.

last_update_details: Option<RoomModification>

Details about the last update to the room.

variant: Option<i32>

The variant / mode of the application being played; can be any integer value, or left blank.

participants: Option<Vec<RoomParticipant>>

The participants involved in the room, along with their statuses. Includes participants who have left or declined invitations.

room_id: Option<String>

Globally unique ID for a room.

application_id: Option<String>

The ID of the application being played.

inviter_id: Option<String>

The ID of the participant that invited the user to the room. Not set if the user was not invited to the room.