Struct google_games1::RoomLeaveRequest

pub struct RoomLeaveRequest {
    pub kind: Option<String>,
    pub reason: Option<String>,
    pub leave_diagnostics: Option<RoomLeaveDiagnostics>,
}

This is a JSON template for a leave room request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

leave rooms (request)

Fields
kind: Option<String>

Uniquely identifies the type of this resource. Value is always the fixed string games#roomLeaveRequest.

reason: Option<String>

Reason for leaving the match. Possible values are:
- "PLAYER_LEFT" - The player chose to leave the room.. - "GAME_LEFT" - The game chose to remove the player from the room. - "REALTIME_ABANDONED" - The player switched to another application and abandoned the room. - "REALTIME_PEER_CONNECTION_FAILURE" - The client was unable to establish a connection to other peer(s). - "REALTIME_SERVER_CONNECTION_FAILURE" - The client was unable to communicate with the server. - "REALTIME_SERVER_ERROR" - The client received an error response when it tried to communicate with the server. - "REALTIME_TIMEOUT" - The client timed out while waiting for a room. - "REALTIME_CLIENT_DISCONNECTING" - The client disconnects without first calling Leave. - "REALTIME_SIGN_OUT" - The user signed out of G+ while in the room. - "REALTIME_GAME_CRASHED" - The game crashed. - "REALTIME_ROOM_SERVICE_CRASHED" - RoomAndroidService crashed. - "REALTIME_DIFFERENT_CLIENT_ROOM_OPERATION" - Another client is trying to enter a room. - "REALTIME_SAME_CLIENT_ROOM_OPERATION" - The same client is trying to enter a new room.

leave_diagnostics: Option<RoomLeaveDiagnostics>

Diagnostics for a player leaving the room.